Update CMake/common, fix link libraries for target export

@jhernando @segalindo @jjgarcia

See merge request !3

+++ fires/Aggregate.h
@@ -191,7 +191,8 @@ namespace fires
       {
         if (( *f ).type == FeatureAggregate::MEAN )
         {
-          ( *f ).aggregate->divide( aggregatedObj, ( *f ).label, objs.size( ));
+          ( *f ).aggregate->divide( aggregatedObj, ( *f ).label,
+				    ( unsigned int ) objs.size( ));
         }
       }
